Roll back sncloud -> SambaNova Cloud
Per Rodrigo M., there are too many places in the
broader code base that rely on the API names
"sncloud" and "sambastudio" to change globally.
So, instead of changing these names in _code_, I've
made a more cosmetic fix so that it's only a change
in the UI, and the naming is then recast to code names
before any functions are executed.